requirements.txt
pandas
streamlit
openai
faiss-cpu
langchain
python-dotenv